moved structs to types module
