+++ v0_4/src/fund_module.rs
@@ -1,8 +1,8 @@
 imports!();
 
-use super::fund_list::*;
-use super::fund_item::*;
-use super::fund_type::*;
+use crate::types::fund_list::*;
+use crate::types::fund_item::*;
+use crate::types::fund_type::*;
 
 
 /// Deals with storage data about delegators.
