+++ v0_4/src/genesis.rs
@@ -1,7 +1,7 @@
 imports!();
 
-use crate::node_state::*;
-use crate::fund_type::*;
+use crate::types::node_state::*;
+use crate::types::fund_type::*;
 
 use crate::events::*;
 use crate::node_activation::*;
