+++ v0_4/src/lib.rs
@@ -6,13 +6,9 @@
 
 // auxiliaries
 pub mod auction_proxy;
-pub mod bls_key;
-pub mod node_state;
-pub mod fund_item;
-pub mod fund_list;
-pub mod fund_type;
+pub mod types;
 
-use crate::bls_key::*;
+use crate::types::bls_key::*;
 
 // modules
 pub mod events;
