+++ v0_4/src/types/fund_type.rs
