+++ v0_4/src/types/node_state.rs
